added topology refactor
